feat: update flujo_caja model to use Float for monetary values and enhance calcular_periodos function with additional payment frequencies

+++ app/models/flujo_caja.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Date, ForeignKey
+from sqlalchemy import Column, Integer, Date, ForeignKey, Float
 from app.models.user import Base
 from sqlalchemy.orm import relationship
 
@@ -9,8 +9,8 @@ class FlujoCaja(Base):
     id = Column(Integer, primary_key=True)
     numero_cuota = Column(Integer)
     fecha = Column(Date)
-    amortizacion = Column(Integer)
-    interes = Column(Integer)
-    cuota = Column(Integer)
-    saldo = Column(Integer)
+    amortizacion = Column(Float)
+    interes = Column(Float)
+    cuota = Column(Float)
+    saldo = Column(Float)
     bono_id = Column(Integer, ForeignKey("bono.id"))
